refactor(MemberStatus) : MemberStatus enum 및 테스트에서 예외 클래스 경로 수정
